requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
decorator==5.1.1
Django==4.2
docker==6.1.2
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
jmespath==1.0.1
kaitaistruct==0.10
lxml==4.9.2
magic-filter==1.0.9
matplotlib-inline==0.1.6
multidict==6.0.4
outcome==1.2.0
packaging==23.1
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
playwright==1.33.0
prompt-toolkit==3.0.38
Protego==0.2.1
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pydivert==2.1.0
pyee==9.0.4
Pygments==2.15.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
requests==2.29.0
requests-file==1.5.1
rsa==4.9
Scrapy==2.8.0
selenium==4.9.0
selenium-wire==5.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
Telethon==1.28.5
tldextract==3.4.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
wcwidth==0.2.6
websocket-client==1.5.1
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
zstandard==0.21.0

arrow~=1.2.3